restructured GSM/MSP code (won't build fyi)
