+++ MSP432/uartGSM/src/sysbios/makefile
@@ -1,7 +1,7 @@
 vpath % /opt/ti/tirtos_msp43x_2_14_03_28/products/bios_6_42_03_35/packages/ti/sysbios/
 vpath %.c /opt/ti/xdctools_3_31_01_33_core/packages/
 
-CCOPTS    = -mcpu=cortex-m4 -march=armv7e-m -mthumb -mfloat-abi=hard -mfpu=fpv4-sp-d16 -D__MSP432P401R__ -DTARGET_IS_FALCON -Dgcc -D_POSIX_SOURCE -DMSP432WARE -I"/opt/ti/ccsv6/ccs_base/arm/include" -I"/opt/ti/ccsv6/ccs_base/arm/include/CMSIS" -I"/home/david/ccs_workspace/uartGSM" -I"/opt/ti/tirtos_msp43x_2_14_03_28/products/MSPWare_2_00_00_40c/driverlib/inc/CMSIS" -I"/opt/ti/tirtos_msp43x_2_14_03_28/products/MSPWare_2_00_00_40c/driverlib/inc" -I"/opt/ti/tirtos_msp43x_2_14_03_28/products/MSPWare_2_00_00_40c/driverlib/MSP432P4xx" -I"/opt/ti/tirtos_msp43x_2_14_03_28/packages/ti/drivers/wifi/cc3100/Simplelink" -I"/opt/ti/tirtos_msp43x_2_14_03_28/packages/ti/drivers/wifi/cc3100/Simplelink/Include" -I"/opt/ti/ccsv6/tools/compiler/gcc-arm-none-eabi-4_8-2014q3/arm-none-eabi/include" -O2 -ffunction-sections -fdata-sections -g -gstrict-dwarf    -Wunused -Wunknown-pragmas -ffunction-sections -fdata-sections  -mcpu=cortex-m4 -mthumb -mfloat-abi=hard -mfpu=fpv4-sp-d16 -mabi=aapcs -g  -O3 -I/opt/ti/tirtos_msp43x_2_14_03_28/products/bios_6_42_03_35/packages/gnu/targets/arm//libs/install-native/$(GCCTARG)/include -Dfar= -D__DYNAMIC_REENT__   -Dti_sysbios_knl_Clock_stopCheckNext__D=FALSE -Dti_sysbios_knl_Task_minimizeLatency__D=FALSE -Dti_sysbios_family_arm_m3_Hwi_enableException__D=TRUE -Dti_sysbios_family_arm_m3_Hwi_disablePriority__D=32U -Dti_sysbios_family_arm_m3_Hwi_numSparseInterrupts__D=0U
+CCOPTS    = -mcpu=cortex-m4 -march=armv7e-m -mthumb -mfloat-abi=hard -mfpu=fpv4-sp-d16 -D__MSP432P401R__ -DTARGET_IS_FALCON -Dgcc -D_POSIX_SOURCE -DMSP432WARE -I"/opt/ti/ccsv6/ccs_base/arm/include" -I"/opt/ti/ccsv6/ccs_base/arm/include/CMSIS" -I"/home/david/ccs_workspace/uartGSM" -I"/opt/ti/tirtos_msp43x_2_14_03_28/products/MSPWare_2_00_00_40c/driverlib/inc/CMSIS" -I"/opt/ti/tirtos_msp43x_2_14_03_28/products/MSPWare_2_00_00_40c/driverlib/inc" -I"/opt/ti/tirtos_msp43x_2_14_03_28/products/MSPWare_2_00_00_40c/driverlib/MSP432P4xx" -I"/opt/ti/tirtos_msp43x_2_14_03_28/packages/ti/drivers/wifi/cc3100/Simplelink" -I"/opt/ti/tirtos_msp43x_2_14_03_28/packages/ti/drivers/wifi/cc3100/Simplelink/Include" -I"/opt/ti/ccsv6/tools/compiler/gcc-arm-none-eabi-4_8-2014q3/arm-none-eabi/include" -ffunction-sections -fdata-sections -g -gdwarf-3 -gstrict-dwarf    -Wunused -Wunknown-pragmas -ffunction-sections -fdata-sections  -mcpu=cortex-m4 -mthumb -mfloat-abi=hard -mfpu=fpv4-sp-d16 -mabi=aapcs -g  -O3 -I/opt/ti/tirtos_msp43x_2_14_03_28/products/bios_6_42_03_35/packages/gnu/targets/arm//libs/install-native/$(GCCTARG)/include -Dfar= -D__DYNAMIC_REENT__   -Dti_sysbios_knl_Clock_stopCheckNext__D=FALSE -Dti_sysbios_knl_Task_minimizeLatency__D=FALSE -Dti_sysbios_family_arm_m3_Hwi_enableException__D=TRUE -Dti_sysbios_family_arm_m3_Hwi_disablePriority__D=32U -Dti_sysbios_family_arm_m3_Hwi_numSparseInterrupts__D=0U
 
 XDC_ROOT  = /opt/ti/xdctools_3_31_01_33_core/packages/
