+++ MSP432/uartGSM/uartecho.cfg
@@ -106,9 +106,9 @@ Task.numPriorities = 4;
 
 /* ================ Task configuration ================ */
 var task0Params = new Task.Params();
-task0Params.instance.name = "echo";
+task0Params.instance.name = "gsm";
 task0Params.stackSize = 0x300;
-Program.global.echo = Task.create("&echoFxn", task0Params);
+Program.global.gsm = Task.create("&gsmTask", task0Params);
 
 /* ================ Logging configuration ================ */
 var LoggingSetup = xdc.useModule('ti.uia.sysbios.LoggingSetup');
